requirements.txt
# keiba-vpn — Web Server + Data Platform
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Scraping
beautifulsoup4>=4.12.0
pymupdf>=1.24.0
playwright>=1.40.0
schedule>=1.2.0

# ML Pipeline
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
optuna>=3.0.0
joblib>=1.3.0

# MLflow
mlflow>=2.12.0

# GCS
google-cloud-storage>=2.0.0
